feat: use trait
