+++ src/libs/utils/src/lib.rs
@@ -3,6 +3,7 @@
 mod doc;
 mod json;
 mod serializers;
+mod types;
 pub mod with;
 
 pub use doc::*;
@@ -12,3 +13,4 @@ pub use crate::serializers::types::{
     DocDataBigInt, DocDataPrincipal, DocDataUint8Array, JsonDataBigInt, JsonDataPrincipal,
     JsonDataUint8Array,
 };
+pub use crate::types::{FromJsonData, IntoJsonData};
