+++ src/sputnik/src/lib.rs
@@ -6,6 +6,7 @@ mod hooks;
 mod js;
 mod polyfill;
 mod state;
+mod functions;
 
 // Developer's custom serverless functions endpoints. Generated by build.rs.
 mod generated;
